Extract plan_mode_helpers.rs from turn.rs (parent #680)
Move the plan-mode pure helpers out of turn.rs into a new
src/agent/loop_run/plan_mode_helpers.rs sibling module:

- should_materialize_plan_after_timeout (pub(super))
- should_materialize_plan_after_tool_call_format_error (pub(super))
- assistant_model_for_mode (pub(super))
- should_fallback_plan_model_after_timeout (pub(super))
- plan_file_alias (pub(super))
- prune_plan_mode_messages (pub(super))
- is_plan_mode_only_system_note (private)

No facade re-export (DR3-001). Callers updated:
- scaffold_pipeline.rs: switches the three `should_*` imports from
  `super::turn` to `super::plan_mode_helpers`.
- commands.rs: switches `prune_plan_mode_messages` from
  `super::turn::` to `super::plan_mode_helpers::`.
- turn.rs: re-imports `assistant_model_for_mode` + `plan_file_alias`
  for production use, plus #[cfg(test)] re-imports for test-only
  references in nested test modules.

turn.rs: 26,919 -> 26,846 (-73 LOC)
Cumulative from 39,489: -12,643 LOC (-32.0%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
